Restart #2
After the mysterious corruption of the project file and deciding that I
can get it working better… restarted from scratch and got the basic
functionality of the interface working.

Still need to figure out the method in which it will select and/or
write to the text file via menu selection and generate buttons.